Moved ApiSettings.kt to the sdk  and left the intreface for it in rtl
