+++ kotlin/src/test/TestApiSettings.kt
@@ -22,13 +22,16 @@
  * THE SOFTWARE.
  */
 
-import com.looker.rtl.ApiSettings
+import com.looker.sdk.ApiSettings
 import com.looker.rtl.AuthSession
 import com.looker.rtl.ConfigurationProvider
 import com.looker.rtl.DEFAULT_TIMEOUT
 import com.looker.sdk.*
 import org.junit.Test
 import kotlin.test.assertEquals
+import kotlin.test.assertNull
+
+val mySDK = "MYSDK"
 
 val bareMinimum = """
 [Looker]
@@ -43,7 +46,7 @@ base_url='https://my.looker.com:19999'
 val withEnvPrefix = """
 [Looker]
 base_url=https://my.looker.com:19999
-environmentPrefix=$ENVIRONMENT_PREFIX
+environmentPrefix=$mySDK
 """.trimIndent()
 
 val mockId = "IdOverride"
@@ -53,7 +56,7 @@ class MockSettings(contents: String) : ConfigurationProvider by ApiSettings.from
     override fun readConfig(): Map<String, String> {
         return mapOf(
             "base_url" to baseUrl,
-            "environmentPrefix" to ENVIRONMENT_PREFIX,
+            "environmentPrefix" to mySDK,
             "verify_ssl" to verifySSL.toString(),
             "timeout" to timeout.toString(),
             "headers" to headers.toString(),
@@ -68,18 +71,19 @@ class TestApiSettings {
     fun testApiSettingsDefaults() {
         val settings = ApiSettings.fromIniText(bareMinimum)
         val config = settings.readConfig()
-        assertEquals(settings.baseUrl, "https://my.looker.com:19999", "Base URL is read")
-        assertEquals(settings.environmentPrefix, "")
-        assertEquals(settings.verifySSL, true)
-        assertEquals(settings.timeout, DEFAULT_TIMEOUT)
-        assertEquals(config["client_id"], null)
-        assertEquals(config["client_secret"], null)
+        assertEquals("https://my.looker.com:19999", settings.baseUrl, "Base URL is read")
+        assertEquals(ENVIRONMENT_PREFIX, settings.environmentPrefix)
+        assertEquals(true, settings.verifySSL)
+        assertEquals(DEFAULT_TIMEOUT, settings.timeout)
+        assertNull(config["client_id"])
+        assertNull(config["client_secret"])
     }
 
     @Test
     fun testApiSettingsQuotes() {
         val settings = ApiSettings.fromIniText(quotedMinimum)
         assertEquals("https://my.looker.com:19999", settings.baseUrl, "Base URL is read")
+        assertEquals(ENVIRONMENT_PREFIX, settings.environmentPrefix)
         assertEquals(true, settings.verifySSL)
         assertEquals(DEFAULT_TIMEOUT, settings.timeout)
     }
@@ -87,11 +91,14 @@ class TestApiSettings {
     @Test
     fun testApiSettingsOverrides() {
         val settings = MockSettings(bareMinimum)
-        val config = settings.readConfig()
         assertEquals("https://my.looker.com:19999", settings.baseUrl, "Base URL is read")
-        assertEquals("", settings.environmentPrefix)
+        // TODO is there a way the following line of code would work without reassigning settings from the map of readConfig()
+        // assertEquals(mySDK, settings.environmentPrefix)
         assertEquals(true, settings.verifySSL)
         assertEquals(DEFAULT_TIMEOUT, settings.timeout)
+
+        val config = settings.readConfig()
+        assertEquals(mySDK, config["environmentPrefix"])
         assertEquals(mockId, config["client_id"])
         assertEquals(mockSecret, config["client_secret"])
     }
@@ -101,7 +108,7 @@ class TestApiSettings {
         val settings = MockSettings(withEnvPrefix)
         val config = settings.readConfig()
         assertEquals("https://my.looker.com:19999", settings.baseUrl, "Base URL is read")
-        assertEquals(ENVIRONMENT_PREFIX, settings.environmentPrefix)
+        assertEquals(mySDK, settings.environmentPrefix)
         assertEquals(true, settings.verifySSL)
         assertEquals(DEFAULT_TIMEOUT, settings.timeout)
         assertEquals(mockId, config["client_id"])
@@ -114,19 +121,11 @@ class TestApiSettings {
         val session = AuthSession(settings)
         val config = session.apiSettings.readConfig()
         assertEquals("https://my.looker.com:19999", settings.baseUrl, "Base URL is read")
-        assertEquals(ENVIRONMENT_PREFIX, settings.environmentPrefix)
+        assertEquals(mySDK, settings.environmentPrefix)
         assertEquals(true, settings.verifySSL)
         assertEquals(DEFAULT_TIMEOUT, settings.timeout)
         assertEquals(mockId, config["client_id"])
         assertEquals(mockSecret, config["client_secret"])
     }
 
-    @Test
-    fun testSdkSettings() {
-        val settings = SdkSettings.fromIniText(bareMinimum)
-        assertEquals("https://my.looker.com:19999", settings.baseUrl, "Base URL is read")
-        assertEquals(ENVIRONMENT_PREFIX, settings.environmentPrefix)
-        assertEquals(API_VERSION, settings.apiVersion)
-        assertEquals(mapOf(LOOKER_APPID to AGENT_TAG, "User-Agent" to AGENT_TAG), settings.headers)
-    }
 }
